pyproject.toml
[project]
name = "roza-jewerly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-parler>=2.3",
    "django-parler-rest>=2.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "environ>=1.0",
    "factory-boy>=3.3.3",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "uvicorn>=0.38.0",
]
